win32 console utf8 issue fix, use print as log

+++ 3rd/kcp/ikcp.h
@@ -341,10 +341,10 @@ void ikcp_setoutput(ikcpcb *kcp, int (*output)(const char *buf, int len,
 	ikcpcb *kcp, void *user));
 
 // user/upper level recv: returns size, returns below zero for EAGAIN
-int ikcp_recv(ikcpcb *kcp, char *buffer, int len);
+int ikcp_recv(ikcpcb *kcp, char *buffer, size_t len);
 
 // user/upper level send, returns below zero for error
-int ikcp_send(ikcpcb *kcp, const char *buffer, int len);
+int ikcp_send(ikcpcb *kcp, const char *buffer, size_t len);
 
 // update state (call it repeatedly, every 10ms-100ms), or you can ask 
 // ikcp_check when to call it again (without ikcp_input/_send calling).
@@ -361,13 +361,13 @@ void ikcp_update(ikcpcb *kcp, IUINT32 current);
 IUINT32 ikcp_check(const ikcpcb *kcp, IUINT32 current);
 
 // when you received a low level packet (eg. UDP packet), call it
-int ikcp_input(ikcpcb *kcp, const char *data, long size);
+int ikcp_input(ikcpcb *kcp, const char *data, size_t size);
 
 // flush pending data
 void ikcp_flush(ikcpcb *kcp);
 
 // check the size of next message in the recv queue
-int ikcp_peeksize(const ikcpcb *kcp);
+size_t ikcp_peeksize(const ikcpcb *kcp);
 
 // change MTU size, default is 1400
 int ikcp_setmtu(ikcpcb *kcp, int mtu);
